+++ app/(tasks)/tasks/[id]/_components/index.ts
@@ -1,2 +1,1 @@
-export * from './ApiTabs';
-export * from './LevelCards';
+export * from './Level';
