--- app/(tasks)/tasks/[id]/_contexts/taskSettings/index.ts
@@ -1,5 +1,0 @@
-'use client';
-
-export * from './TaskSettingsContext';
-export * from './TaskSettingsProvider';
-export * from './useTaskSettings';
