--- app/(tasks)/tasks/[id]/_contexts/taskSettings/useTaskSettings.ts
@@ -1,5 +1,0 @@
-import { use } from 'react';
-
-import { TaskSettingsContext } from './TaskSettingsContext';
-
-export const useTaskSettings = () => use(TaskSettingsContext);
