+++ app/(tasks)/tasks/[id]/_helpers/getTaskSettingsCookieValue.ts
@@ -4,7 +4,7 @@ import { cookies } from 'next/headers';
 
 import { COOKIES } from '@/constants';
 
-import type { TaskSettingsCookieValue } from '../_contexts/taskSettings';
+import type { TaskSettingsCookieValue } from '../_types';
 
 export const getTaskSettingsCookieValue = async () => {
   const cookieStore = await cookies();
