+++ app/(tasks)/tasks/[id]/_types/index.ts
@@ -34,3 +34,8 @@ export interface TaskContent {
   };
   title: MessagePath;
 }
+
+export interface TaskSettingsCookieValue {
+  api: ApiType;
+  level: LevelName;
+}
